Converted useMenu hook to typescript
